MM-422 et al bust browser cache
Since we can't find where the bad JSON escaping (or lack there of) is happening, I'm thinking it may be due to browser cached old versions. This changes the brunch build to name the app.js and libs.js files with their md5 checksum and use that in the index.html.
Also resolved several security issues in node components.
Also fixed proptypes warning on issue and article checkboxes